fix: Fix multiple reliability issues with docker (#172)
This PR fixes a host of issuess with the docker executor. Additionally,
it fixes the integration tests on CI, dramatically increasing test
coverage. Tests now also use a much smaller and faster image.

A non-exhaustive list:
* Executor now packs symlinks correctly
* Executor will now copy .git/, even if it's in an ignore file
* Executor will now forward build errors if the dockerfile has any
* Executor now uses a shared, lazy client to docker, instead of one per
agent
* Docker images are now better named, and tagged with the first 'part'
of the agent uuid

Additionally adds an entry in the Justfile to run the tests in docker
(useful when running on a mac)