refactor: Add support for experimenting with multiple agents (#231)
We want to be able to experiment with different kinds of agents, without
affecting our main users. With this an agent model can be selected in
the config (defaulting to what we think works best). Later, we could
i.e. add a development agent, experiment with multi-agent setups, etc,
without affecting our users.