feat: Agent sessions with multiple agents in a single chat (#245)
**This does NOT set the new PlanAct flow as the default yet.**

Currently, agents are overly enthusiastic in solving the issues by
coding right away. That's certainly not intended. We can't really prompt
out of this, the agent would then just stop more often and ask for
feedback.

I have a simple idea in mind to create a plan -> act agent setup. The
planner either responds to the question, researches, and if coding is
needed, first asks for confirmation before starting the coding task.
What makes this `simple` is that there is only one active agent at a
given time. So once the coding agent starts, the active agent in the
session is the coding agent. Once done, it switches back to the initial
agent. Both agents operate on the same executor with their own message
history. During the chat, both agents remain active and run as a
'singleton'.

There's several things this PR intends to tackle architecturally:
* Instead of loose back and forth; an active chat has a 'session' in the
backend. A session can send updates on the session (instead of the 5 ish
updates that are present right now).
* Give structure to managing agents. The session is responsible for
encapsulating all the things the agent needs *in context of the chat*.
* A lot of code can be cleaned up or made more reusable in `v1`

Additionally, this PR should make it easier to build different agent
setups as well.

While I'm at it, let's also question how the 'RunningAgent' fits in
here, and if its needed at all.


```mermaid
sequenceDiagram

  actor U as User
  participant CH as Command Handler
  U->>CH: "How would you implement running multiple agents?"
  CH->>+Session: Create a session 
  create actor Agent
  Session->>+Agent: Create initial agent
  Agent-->>U: "Here is the plan, would you like me to build it?"
  U->>Agent: "Ok!"
  deactivate Agent
  create actor CodingAgent
  Agent->>+CodingAgent: "Please code what the user just said"
  CodingAgent-->>U: Ok!
  loop [coding with tools]
    CodingAgent-->U: Updates
  end
  CodingAgent-->>U: It works!
  deactivate CodingAgent
  deactivate Session
```