refactor: Simplify db code (#339)
commit 4c7c6cfc5ce58443fd3627be092af802c6e9ca64

    refactor: simplify db init code

commit 8d4146614c4f5c422302a62ec18f31b45c16f2a7

    chore: remove unnecessary Arc wrappers around DB
    
A fairly common pattern in types which implement clients (db, web, etc.)
is to have an inner type which is Arc wrapped. Both LanceDB and Redb are
    such types. This commit replaces the unnecessary Arc wrappers around
these types in the storage module and all the places where they are used
    with a simple owned type which is cloned when needed.

---

Note: this is something I noticed when taking a bit of a quick read
through the code base. I haven't been able to get kwaak working due to
https://github.com/bosun-ai/kwaak/issues/338, so unsure if this is
working correctly. Slapping the PR up regardless. Also, I do note that
the db types carry a bit of config data around which is not Arc wrapped.
This seems intuitively like it's small enough that it's unlikely to be a
performance burden (order of magnitude 100s of bytes), and not data
which is expected to be mutated by multiple code paths (there's no Mutex
on the db types), so this data seems reasonable to clone when necessary.

Another approach to this might be to not have the statics and instead
just creating the db values once, passing them between methods and
cloning as needed. I haven't looked fully at the usage patterns to
understand whether this approach would be appropriate, but it is
something that generally works better than using global state like this.