chore(evaluations): Evaluate query pipeline with ragas (#409)
Evaluate query pipelines with ragas. Add an input/output file, and it
will run all the questions, recording the outputs to a ragas compatible
dataset. The dataset can then be evaluated in a python notebook,
recorded to HF, or something else.

Includes an example notebook.


![image](https://github.com/user-attachments/assets/029e966d-9d3e-41a6-8bd1-d0a503dfd128)

We can use this later to evaluate changes in Kwaak's RAG.