feat: Edit with self correcting patches (#441)
The problem: Line editing is unreliable, full writes work but can have
side effects and devour tokens.

Main challenge is that LLMs suck at counting reliably. This solution
removes the need for counting and builds on the assumption that there is
plenty of patch data an LLM trained on.

But, likely the numbers are still close, which helps us deal with any
ambiguity in the patch. Bonus is that with this we can support multiple
edits in a file with a single LLM call and minimal tokens.

How it works:
* First parse the patch into a loosy format we can work with
* Iterate over the lines in the target file, if there is a match with a
hunk, create a candidate
* If that line matches with the next line for any current candidate,
keep the candidate, otherwise ditch
* Finally, we can recalculate the source and dest lines and ranges, and
re-render the patch.

The algorithm supports ambiguity (multiple hunks matching). Still needs
to be dealt with.

TODO:
- [x] Deal with ambiguity
- [x] More tests for multiple hunks
- [x] Test with diffy if the regenerated patches are correct
- [x] Add it all to the tool
- [x] Run some evals

If the evals pass we'll keep it in. After we run a set of SWE, if we see
no issues, as far as I'm concerned we can make it the default (and write
a cool blog article).

Aside, it should also be possible to Cow/borrow most of the inner values
and have this be near zero copy. The way it will be used however it's no
where near a bottleneck. Cool exercise though.